pyproject.toml
[project]
name = "ai-sdgs-scoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "huggingface-hub>=1.6.0",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-qwq>=0.3.4",
    "pymupdf4llm>=1.27.2.1",
    "pytest>=9.0.2",
    "sentence-transformers>=5.2.3",
    "supabase==2.15.2",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]
